feat: 여행 계획 컴포넌트 UI/UX 개선 및 코드 정리
